Sat, Sep 16, 2023, 12:26 PM -03:00

+++ cs50/aula4/filter/helpers.c
@@ -148,9 +148,9 @@ void edges(int height, int width, RGBTRIPLE image[height][width])
                     }
                 }
             }
-            image[h][w].rgbtBlue = sqrt(pow(sumgx.rgbtBlue, 2) + pow(sumgy.rgbtBlue, 2)); //Ver raiz quadrada e elevado
-            image[h][w].rgbtGreen = sqrt(pow(sumgx.rgbtGreen, 2) + pow(sumgy.rgbtGreen, 2));
-            image[h][w].rgbtRed = sqrt(pow(sumgx.rgbtRed, 2) + pow(sumgy.rgbtRed, 2));
+            image[h][w].rgbtBlue = (int) sqrt(pow(sumgx.rgbtBlue, 2) + pow(sumgy.rgbtBlue, 2)); //Ver raiz quadrada e elevado
+            image[h][w].rgbtGreen = (int) sqrt(pow(sumgx.rgbtGreen, 2) + pow(sumgy.rgbtGreen, 2));
+            image[h][w].rgbtRed = (int) sqrt(pow(sumgx.rgbtRed, 2) + pow(sumgy.rgbtRed, 2));
         }
     }
     return;
